CE-3 changes
